% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricCalc_condiVariance.R
\name{applyVar}
\alias{applyVar}
\title{Calculate conditional variance}
\usage{
applyVar(X, margin = 1)
}
\arguments{
\item{X}{matrix of data, NA if not in the inclusion condition}

\item{margin}{which margin to apply function, 1 if rows are genes/probes}
}
\value{
marginal variance
}
\description{
Define variance metric as average squared distance from median to each point greater than the median
}
