% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricCalc_condiVariance.R
\name{applyVar_groups}
\alias{applyVar_groups}
\title{Conditional variance pooled by group}
\usage{
applyVar_groups(X, g, margin = 1)
}
\arguments{
\item{X}{matrix of data, NA if not in the inclusion condition}

\item{g}{group labels, 2 levels}

\item{margin}{which margin to apply function, 1 if rows are genes/probes}
}
\value{
Conditional variance for each gene pooled by group
}
\description{
Define variance metric as average squared distance from median to each point greater than the median.
}
