% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_sensitivity.R
\name{calc_count_cutoff}
\alias{calc_count_cutoff}
\title{Count needed to call signal}
\usage{
calc_count_cutoff(
  postL_cutoff,
  useparEst_itr_out,
  parEst_itr_out = NULL,
  gridResolution = 500,
  alphai_beta_g = NULL,
  sigma2_1g = NULL,
  mu_0i = NULL,
  sigma2_0i = NULL,
  pi_prior = NULL
)
}
\arguments{
\item{postL_cutoff}{cutoff used to binarize postL}

\item{useparEst_itr_out}{boolean indicator to use output object from \code{parEst_itr} function}

\item{parEst_itr_out}{if \code{useparEst_itr_out} is T, then provide the output object from \code{parEst_itr} function}

\item{gridResolution}{number of grid nodes to generate for the step function for posterior likelihood}

\item{alphai_beta_g}{mandatory parameter if \code{useparEst_itr_out} is F, will be ignored otherwise. Matrix of signal mean, after log normal approximation. Same dimensions as \code{Y_mtx}.}

\item{sigma2_1g}{mandatory parameter if \code{useparEst_itr_out} is F, will be ignored otherwise. Matrix of signal variance, after log normal approximation. Same dimensions as \code{Y_mtx}.}

\item{mu_0i}{mandatory parameter if \code{useparEst_itr_out} is F, will be ignored otherwise. Matrix of background mean. Same dimensions as \code{Y_mtx}.}

\item{sigma2_0i}{mandatory parameter if \code{useparEst_itr_out} is F, will be ignored otherwise. Matrix of background variance. Same dimensions as \code{Y_mtx}.}

\item{pi_prior}{mandatory parameter if \code{useparEst_itr_out} is F, will be ignored otherwise. Matrix of prior likelihood of signal. Same dimensions as \code{Y_mtx}.}
}
\value{
Matrix of posterior likelihood, same dimensions as \code{alphai_beta_g}.
}
\description{
Count needed to reach desired postL for each sample and gene combination
}
