% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodRatioTest.R
\name{calc_S_LR}
\alias{calc_S_LR}
\title{calculate test statistics for LRT}
\usage{
calc_S_LR(Z0, Z1, n0, n1, s0, s1)
}
\arguments{
\item{s1}{}
}
\value{

}
\description{
calculate test statistics for LRT
}
