% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_BG_calling.R
\name{call_BG_wNegCtrl_Zscore}
\alias{call_BG_wNegCtrl_Zscore}
\title{Call background with default methods, Z score}
\usage{
call_BG_wNegCtrl_Zscore(Ymtx, Neg_control, method, binary = NULL)
}
\arguments{
\item{Ymtx}{Matrix of raw counts, gene by sample. Endogenous probes only}

\item{Neg_control}{Matrix of raw counts, gene by sample. Negative control probes only}

\item{method}{ways for calling background, "max" for less than max of all negative controls; "norm" model negative controls with normal distributions; "logNorm" model negative controls with log normal distributions}

\item{binary}{cutoff used binarize Z score. \code{method = "max"} will always return binary labels. If false, the other two method will return p-values for observing each count from sample specific background distribution.}
}
\value{
matrix of binary labels (1 for signal and 0 for background), or Z scores for observing each count from sample specific background distribution. Same dimensions as \code{Ymtx}
}
\description{
call background according to only negative control probes
}
