% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricCalc_condiVariance.R
\name{deltaMedian}
\alias{deltaMedian}
\title{Difference in group medians}
\usage{
deltaMedian(X, g, margin = 1)
}
\arguments{
\item{X}{matrix of data, NA if not in the inclusion condition}

\item{g}{group labels, 2 levels}

\item{margin}{which margin to apply function, 1 if rows are genes/probes}
}
\value{
Difference in group medians, 2nd level - 1st level
}
\description{
Difference in group medians
}
