% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit_wrapper.R
\name{fit_MMD}
\alias{fit_MMD}
\title{Wrapper function to fit GMM}
\usage{
fit_MMD(Y_mtx, maxIter = 20, NegCtrl = NULL, Z_mtx = NULL, verbose = F, ...)
}
\arguments{
\item{Y_mtx}{Matrix of raw counts, gene by sample}

\item{maxIter}{Maximum iterations to fit the model}

\item{NegCtrl}{Matrix of raw counts for negative control probes, gene by sample. Must have the same number of columns as \code{Y_mtx}. Will be ignored if \code{Z_mtx} is provided}

\item{Z_mtx}{Optional matrix of binary labels, same dimensions as \code{Y_mtx}. 0 as in background, and 1 for signal.}

\item{verbose}{Print steps of model fitting if True}

\item{...}{The other provided arguments will go into \code{parEst_itr()}}
}
\value{
a list of objects
\itemize{
\item \code{model_fit} fitted MMD model. Output from \code{parEst_itr()}
\item \code{Z_mtx_adj} Matrix of posterior likelihood after monotonicity adjustment. Ouput from \code{forceMono()}
}
}
\description{
This function calls \code{parEst_itr()} and \code{forceMono()}. It fits the MMD iteratively and by the end calculate a posterior likelihood for each observed count containing signal.
}
