% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodRatioTest.R
\name{p_null_MLE}
\alias{p_null_MLE}
\title{p null MLE, using the known sensitivities}
\usage{
p_null_MLE(Z0, Z1, n0, n1, s0, s1)
}
\arguments{
\item{s1}{}
}
\value{

}
\description{
p null MLE, using the known sensitivities
}
