% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodRatioTest.R
\name{S_LRT_null}
\alias{S_LRT_null}
\title{simulate null distribution for LRT statistics}
\usage{
S_LRT_null(n0, n1, p0, s0, s1, nRep = 1000)
}
\arguments{
\item{nRep}{}
}
\value{

}
\description{
simulate null distribution for LRT statistics
}
