% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_DP.R
\name{sim_DP_sampling}
\alias{sim_DP_sampling}
\title{Function to simulate a set of data, and introduce differential presence}
\usage{
sim_DP_sampling(
  nEndogenous = 800,
  nPosControl = 10,
  nNegControl = 10,
  nSample = 500,
  d0 = 50,
  expressPercent = 1/5,
  p0_mean = 1/5,
  p0_size = 1000,
  seed.par = NULL,
  seed.ZY = NULL,
  nGroupA,
  nDPgenes,
  DP_effectSize,
  batchEffect = 0
)
}
\arguments{
\item{nEndogenous}{number of endogenous probes, default = 800}

\item{nPosControl}{number of positive control probes, default = 10}

\item{nNegControl}{number of negative control probes, default = 10}

\item{nSample}{total number of samples, default = 500}

\item{d0}{prior degree of freedom for simulating signal variance}

\item{expressPercent}{fixed proportion of endogenous probes expressed in each sample}

\item{p0_mean}{for null genes, mean parameter to simulate binomial rate}

\item{p0_size}{for null genes, size parameter to simulate binomial rate}

\item{seed.par}{seed for simulating parameters}

\item{seed.ZY}{seed for simulating Z and Y}

\item{nGroupA}{number of sample for one of the group of interest}

\item{nDPgenes}{number of genes to introduce differential presence, will be split evenly to have higher likelihood to be detected in group A or B}

\item{DP_effectSize}{how much more likely to be present in one group, on top of the baseline expressPercent}

\item{alpha_shift}{introduced confounding between groups of samples. Group A shift downward by alpha_shift/2, group B shift upward by alpha_shift/2}
}
\value{

}
\description{
Different from \code{sim_DP}, this function randomly allocate 100 samples ($\alpha_i$ for each is fixed) into two groups, creating balanced or unbalanced groups. Then introduce DP for a subset of probes as before.
}
