% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_DP.R
\name{sim_DP_wRealData}
\alias{sim_DP_wRealData}
\title{Simulate nCounter samples and selectively introduce differential presence to subset of probes}
\usage{
sim_DP_wRealData(
  model_fit = NULL,
  sampleGroup = NULL,
  baseline_SigPrev = 0.3,
  effectSize = 0.3,
  nPosControl = 10,
  nNegControl = 10,
  nSample = 500,
  d0 = 50,
  expressPercent = 1/5,
  beta_offset = 0,
  seed.par = 0,
  seed.data = 0,
  nDPgenes = 200
)
}
\arguments{
\item{model_fit}{optional input, previously estimated MMD model parameters that contains study of interest}

\item{sampleGroup}{vector same length as total number of sampled used to fit \code{model_fit}. For samples that's not in study of interest will have value NA, and the two groups in the study will have value 0 or 1 to indicate group assignment.}

\item{baseline_SigPrev}{baseline signal prevalence}

\item{effectSize}{target effect size for differential presence, will be added on top of \code{baseline_SigPrev}}

\item{nPosControl}{number of positive control probes, default = 10}

\item{nNegControl}{number of negative control probes, default = 10}

\item{nSample}{total number of samples, default = 500}

\item{d0}{prior degree of freedom for simulating signal variance}

\item{expressPercent}{fixed proportion of endogenous probes expressed in each sample}

\item{beta_offset}{a numerical value to offset all beta_offset, the larger the further away all signals will be from the noise}

\item{seed.par}{seed for simulating parameters, default = 0}

\item{seed.data}{seed for simulating Z and Y, default = 0}

\item{nDPgenes}{number of genes to introduce differential presence, will be split evenly to have higher likelihood to be detected in group A or B}
}
\value{

}
\description{
Simulate nCounter samples and selectively introduce differential presence to subset of probes
}
