% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_null.R
\name{simData}
\alias{simData}
\title{Simulate NanoString nCounter data with from null model}
\usage{
simData(
  nEndogenous = 800,
  nPosControl = 10,
  nNegControl = 5,
  nSample = 500,
  alwaysOnGenes = 0,
  fixedBackground = T,
  d0 = 50,
  alpha_Offset = 2
)
}
\arguments{
\item{nEndogenous}{number of endogenous probes, default = 800}

\item{nPosControl}{number of positive control probes, default = 10}

\item{nNegControl}{number of negative control probes, default = 5}

\item{nSample}{total number of samples, default = 500}

\item{alwaysOnGenes}{for the endogenous probes, set certain amount always expressed in all samples, default = 0}

\item{fixedBackground}{boolean indicator, if true, all samples will share the same background mean and variance parameters}

\item{d0}{prior degree of freedom for simulating signal variance}

\item{alpha_Offset}{a numerical value to offset all alpha_i, the larger the further away all signals will be from the noise}
}
\value{
a list of data and simulated true parameters

\itemize{
\item \code{Y_mtx} a matrix of all observed counts for endogenous probes. Probes by samples
\item \code{Z_mtx} a matrix of true labels for endogenous probes, 0 for noise, 1 for signal, same dimensions as \code{Y_mtx}
\item \code{Neg_control}  a matrix of all observed counts for Negative control probes. Probes by sample
\item \code{Pos_control}  a matrix of all observed counts for Positive control probes. Probes by sample
\item \code{pi_i} vector of simulated sample marginal prior probabilty of observing signal
\item \code{mu_0i} vector simulated background mean
\item \code{sigma_0i} vector of simulate background variance
\item \code{alpha_i} vector of simulated sample contribution to signal mean, offset included
\item \code{beta_g} vector of simulated probe contribution to signal mean
\item \code{alpha_i_beta_g} matrix of signal mean, same dimensions as \code{Y_mtx}
\item \code{sigma2_10} vector of signal variance prior
\item \code{sigma2_1g} vector of simulated signal variance
}
}
\description{
simulate parameter and data, with potentially varying prior probability of signal in all samples
}
