% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricCalc_condiVariance.R
\name{tLikeStat}
\alias{tLikeStat}
\title{T-like statistics}
\usage{
tLikeStat(X, g, margin = 1)
}
\arguments{
\item{X}{matrix of data, NA if not in the inclusion condition}

\item{g}{group labels, 2 levels}

\item{margin}{which margin to apply function, 1 if rows are genes/probes}
}
\value{
T-like statistics
}
\description{
calculated as change in group medians/SE, were SE is calculated with pooled conditional variance and df.
}
